Split SYCL detail tests in hipSYCL and DPC++ directories.
